+++ src/logger.c
@@ -249,7 +249,7 @@ lookup_syslog_facility(const char *facility) {
         { "uucp",   LOG_UUCP },
     };
 
-    for (int i = 0; i < sizeof(facilities) / sizeof(facilities[0]); i++)
+    for (size_t i = 0; i < sizeof(facilities) / sizeof(facilities[0]); i++)
         if(strncasecmp(facilities[i].name, facility, strlen(facility)) == 0)
             return facilities[i].number;
